requirements_dev.txt
pip==8.1.2
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==3.0.4
tox==2.3.1
coverage==4.2
pytest-aiohttp==0.1.3
Sphinx==1.4.8
cryptography==1.5.2
PyYAML==3.12
pytest==3.0.3
aiohttp_validate==0.1.1